Fix the css file
Place the .small-para selector after the
.para selector. Since both have the same level of
specificity, what determines which style gets
applied is which one appears last.

Move the .confirm selector after the .button
selector for the same reasons stated above.

Add the .text selector to the .child selector
to form a chain so that the styles get applied
properly.